% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_photos_roe_data.R
\name{check_photos_roe_data}
\alias{check_photos_roe_data}
\title{Fonction pour checker la présence de photos sur un ouvrage identifié avec son numéro.}
\usage{
check_photos_roe_data(id_ouvrage, session)
}
\arguments{
\item{id_ouvrage}{Numéro de l'ouvrage ROE}

\item{session}{Résultat de la connexion/session à GéObs}
}
\value{
un dataframe avec la présence et le nombre de photos pour l'ouvrage consulté
}
\description{
Fonction pour checker la présence de photos sur un ouvrage identifié avec son numéro.
}
\examples{
\dontrun{

## initier la session geobs
ma_session <- init_geobs_session(
  login = "john.doe@ofb.gouv.fr",
  mdp = "geobs_mon_amour")

## checker pour l'ouvrage "ROE39268"
check_photos_roe_data(id_ouvrages = 39268, session = ma_session)
}

}
