% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bdoe_data.R
\name{get_bdoe_data}
\alias{get_bdoe_data}
\title{Fonction pour télécharger les données du module BDOE de GéObs.}
\usage{
get_bdoe_data(login, mdp, nom_dossier = NULL, lecture = FALSE)
}
\arguments{
\item{login}{texte. l'adresse mail d'accès à l'application GéObs.}

\item{mdp}{texte. le mot de passe associé.}

\item{nom_dossier}{texte. le nom du dossier où seront téléchargées les données.}

\item{lecture}{booléen. si TRUE, lit le fichier zip téléchargé.}
}
\value{
un fichier zip nommé \code{export_csv_total_bd.zip} contenant les données est téléchargé.
 Si le fichier zip est lu, retourne un objet sous forme de liste de dataframe.
}
\description{
Fonction pour télécharger les données du module BDOE de GéObs.
}
\examples{
\dontrun{
get_bdoe_data(login = "john.doe@ofb.gouv.fr",
              mdp = 'geobs_mon_amour',
              nom_dossier = "data_bdoe")

bdoe_data <-
  get_bdoe_data(login = "john.doe@ofb.gouv.fr",
                mdp = 'geobs_mon_amour',
                nom_dossier = "data_bdoe",
                lecture = TRUE)
}
}
