% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CE_BDTOPO_wfs.R
\name{get_CE_BDTOPO_wfs}
\alias{get_CE_BDTOPO_wfs}
\title{Récupérer les Cours d'eau de la BDTOPO IGN sous forme spatiale}
\usage{
get_CE_BDTOPO_wfs(geo)
}
\arguments{
\item{geo}{objet spatial de type 'sf'}
}
\value{
un data.frame au format 'sf'
}
\description{
Récupérer les Cours d'eau de la BDTOPO IGN sous forme spatiale
}
\examples{
\dontrun{
get_CE_BDTOPO_wfs(geo = shape_area)
}

}
