% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classCE_wfs.R
\name{get_classCE_wfs}
\alias{get_classCE_wfs}
\title{Récupérer le classement Liste 1 et 2 des Cours d'eau sous forme spatiale}
\usage{
get_classCE_wfs(liste = "Liste2", geo)
}
\arguments{
\item{liste}{texte. Nom du classement de continuité écologique ('Liste1', 'Liste2')}

\item{geo}{objet spatial de type 'sf'}
}
\value{
un data.frame au format 'sf'
}
\description{
Récupérer le classement Liste 1 et 2 des Cours d'eau sous forme spatiale
}
\examples{
\dontrun{
get_classCE_wfs(liste = 'Liste2', shape_area)
}

}
