% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobs_data.R
\name{get_geobs_data}
\alias{get_geobs_data}
\title{Récupérer les données de la base GEOBS}
\usage{
get_geobs_data(couche, geo)
}
\arguments{
\item{couche}{nom de la couche à charger}

\item{geo}{objet spatial de type 'sf'}
}
\value{
un data.frame au format 'sf'
}
\description{
Récupérer les données de la base GEOBS
}
\examples{
\dontrun{
get_geobs_data("ROE_MONDE", shape_area)
}
}
