% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobs_data_flux.R
\name{get_geobs_data_flux}
\alias{get_geobs_data_flux}
\title{Récupérer les données de la base GEOBS à partir du flux BRGM}
\usage{
get_geobs_data_flux(couche, geo)
}
\arguments{
\item{couche}{nom de la couche à charger}

\item{geo}{objet spatial de type 'sf'}
}
\value{
un data.frame au format 'sf'
}
\description{
Récupérer les données de la base GEOBS à partir du flux BRGM
}
\examples{
\dontrun{
get_geobs_data_flux("ROE_MONDE", shape_area)
}
}
