% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobs_wfs_attributs.R
\name{get_geobs_wfs_attributs}
\alias{get_geobs_wfs_attributs}
\title{Récupérer les noms des attributs d'une couche geobs sur le flux WFS}
\usage{
get_geobs_wfs_attributs(nom_couche)
}
\arguments{
\item{nom_couche}{texte. nom de la couche}
}
\value{
un vecteur de noms des attributs de la couche interrogée.
}
\description{
Récupérer les noms des attributs d'une couche geobs sur le flux WFS
}
\examples{
\dontrun{
get_geobs_wfs_attributs("ROE_MONDE")
}
}
