% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobs_wfs_couches.R
\name{get_geobs_wfs_couches}
\alias{get_geobs_wfs_couches}
\title{Récupérer le nom des couches disponibles dans le flux WFS Geobs}
\usage{
get_geobs_wfs_couches()
}
\value{
un vecteur de noms de couches disponibles.
}
\description{
Récupérer le nom des couches disponibles dans le flux WFS Geobs
}
\examples{
\dontrun{
get_geobs_wfs_couches()
}
}
