% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_geobs_session.R
\name{init_geobs_session}
\alias{init_geobs_session}
\title{Fonction pour initier une session de connexion à Geobs}
\usage{
init_geobs_session(login, mdp)
}
\arguments{
\item{login}{texte. Login (email) compte Geobs}

\item{mdp}{texte. Mot de passe compte Geobs}
}
\description{
Initie une connexion à l'application Géobs.
}
\examples{
\dontrun{
init_geobs_session(login, mdp)
}

}
